Add config option: ignore-defined-content
Creates a config setting that causes restyle-define to not expect any captured content.

When captures are expected, every call to restyle define invokes
-restyle--placeholder-name which invokes the checksum function in js. At
scale, this has a measurable performance impact because of the frequency
of use even though most uses of restyle-define do not receive content
blocks.

Ideally we'd use Sass's content-exists() function but it doesn't work in
libSass: https://github.com/sass/libsass/issues/2842. Once that bug is
fixed, we can change restyle-define to just check if the restyle-define mixin was
passed any content and intelligently avoid calling
-restyle--placeholder-name.